BioInfo/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
elastic-transport==8.4.0
elasticsearch==8.9.0
elasticsearch-dsl==7.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
unicodecsv==0.14.1
urllib3==1.26.16
